# ergo-api/web/openapi.yaml
openapi: "3.0.2"
info:
  version: "6.0.2"
  title: Ergo Node API (Rust)
  description: |
    OpenAPI spec inherited from the Scala reference node. The Rust
    node implements most of the Scala wire surface plus a small
    Rust-exclusive operator overlay. This page is the authoritative
    map of what this build serves: the sections below enumerate every
    route that returns **404**, every config-gated route, and every
    Rust-exclusive addition. Anything not called out here matches the
    Scala wire shape (see **Everything else**).

    ## Not implemented in this build

    These paths return **404** under default config. Tracked-but-
    deferred items are listed with their status; off-spec items
    are not on our roadmap.

    **Scan API (not implemented — no equivalent in this build):**

    - `/scan/*` (all 9 routes — register / unspentBoxes / addBox /
      p2sRule / etc.)

    **Script execution (deferred — `ergo-sigma` exposes the
    primitives, no REST wrapper yet):**

    - `/script/p2sAddress`, `/script/p2shAddress`,
      `/script/addressToTree/{address}`,
      `/script/addressToBytes/{address}`, `/script/executeWithContext`

    **Tracked / planned (the underlying logic exists; the REST
    surface is not yet mounted):**

    - `/nipopow/popowHeaderById/{headerId}`,
      `/nipopow/popowHeaderByHeight/{height}`,
      `/nipopow/proof/{minChainLength}/{suffixLength}`,
      `/nipopow/proof/{minChainLength}/{suffixLength}/{headerId}`
...

...

...

...

...

...

...

...

...

...
    fixtures.
...
servers:
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: api_key
...
paths:
  /blocks:
    get:
      summary: Get an array of header ids (hex encoded) for the given range of blockchain block heights.
        Returns a page of the whole list starting from `offset` and containing `limit` items.
      operationId: getHeaderIds
      tags:
        - blocks
      parameters:
        - in: query
          name: limit
          required: false
          description: The number of items in list to return
          schema:
            type: integer
            format: int32
            minimum: 10
            maximum: 100
            default: 50
        - in: query
          name: offset
          required: false
          description: The first block height to include in the list
          schema:
            type: integer
            format: int32
            default: 0
      responses:
        '200':
          description: Array of header ids
          content:
            application/json:
              schema:
                type: array
                description: Array of header ids
                items:
                  type: string
                  example: '8b7ae20a4acd23e3f1bf38671ce97103ad96d8f1c780b5e5e865e4873ae16337'
                example: ['8b7ae20a4acd23e3f1bf38671ce97103ad96d8f1c780b5e5e865e4873ae16337']
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    post:
      summary: Send a mined block
      operationId: sendMinedBlock
      tags:
        - blocks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FullBlock'
      responses:
        '200':
          description: Block is valid
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /blocks/at/{blockHeight}:
    get:
      summary: Get header ids at the given height
      operationId: getFullBlockAt
      tags:
        - blocks
      parameters:
        - in: path
          name: blockHeight
          required: true
          description: Height of a block to retrieve header ids
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Array of header ids
          content:
            application/json:
              schema:
                type: array
                description: Array of header ids
                example: ['8b7ae20a4acd23e3f1bf38671ce97103ad96d8f1c780b5e5e865e4873ae16337']
                items:
                  type: string
                  example: '8b7ae20a4acd23e3f1bf38671ce97103ad96d8f1c780b5e5e865e4873ae16337'
        '404':
          description: Blocks at this height doesn't exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /blocks/chainSlice:
    get:
      summary: Get headers in a specified range of heights
      operationId: getChainSlice
      tags:
        - blocks
      parameters:
        - in: query
          name: fromHeight
          required: false
          description: Min header height (start of the range)
          schema:
            type: integer
            format: int32
            default: 0
        - in: query
          name: toHeight
          required: false
          description: Max header height of the range (last header height then omitted)
          schema:
            type: integer
            format: int32
            default: -1
      responses:
        '200':
          description: Array of headers
          content:
            application/json:
              schema:
                type: array
                description: Array of headers
                items:
                  $ref: '#/components/schemas/BlockHeader'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /blocks/{headerId}:
    get:
      summary: Get the full block info by a given header id
      operationId: getFullBlockById
      tags:
        - blocks
      parameters:
        # todo move out after full support aos3.0 was added to swagger-codegen
        - in: path
          name: headerId
          required: true
          description: ID of the header the wanted block
          schema:
            type: string
      responses:
        '200':
          description: Block object representing the full block data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullBlock'
        '404':
          description: Block with this id doesn't exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /blocks/headerIds:
    post:
      summary: Get full blocks by given header ids
      operationId: getFullBlockByIds
      tags:
        - blocks
      requestBody:
          required: true
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
      responses:
        '200':
          description: Full blocks corresponding to ids provided
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FullBlock'
        '404':
          description: No block exist for every id provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /blocks/{headerId}/header:
    get:
      summary: Get the block header info by a given header id
      operationId: getBlockHeaderById
      tags:
        - blocks
      parameters:
        # todo move out after full support aos3.0 was added to swagger-codegen
        - in: path
          name: headerId
          required: true
          description: ID of a wanted block header
          schema:
            type: string
      responses:
        '200':
          description: Block header object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockHeader'
        '404':
          description: Block with this id doesn't exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /blocks/{headerId}/transactions:
    get:
      summary: Get the block transactions info by a given signature
      operationId: getBlockTransactionsById
      tags:
        - blocks
      parameters:
        # todo move out after full support aos3.0 was added to swagger-codegen
        - in: path
          name: headerId
          required: true
          description: ID of a wanted block transactions
          schema:
            type: string
      responses:
        '200':
          description: Block transaction object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockTransactions'
        '404':
          description: Block with this id doesn't exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /blocks/{headerId}/proofFor/{txId}:
    get:
      summary: Get Merkle proof for transaction
      operationId: getProofForTx
      tags:
        - blocks
      parameters:
        - in: path
          name: headerId
          required: true
          description: ID of a wanted block transactions
          schema:
            type: string
        - in: path
          name: txId
          required: true
          description: ID of a wanted transaction
          schema:
            type: string
      responses:
        '200':
          description: Merkle proof object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerkleProof'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /blocks/lastHeaders/{count}:
    get:
      summary: Get the last headers objects
      operationId: getLastHeaders
      tags:
        - blocks
      parameters:
        # todo move out after full support aos3.0 was added to swagger-codegen
        - in: path
          name: count
          required: true
          description: a number of block headers to return
          schema:
            type: number
      responses:
        '200':
          description: Array of block headers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BlockHeader'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /peers/all:
    get:
      summary: Get all known peers
      operationId: getAllPeers
      tags:
        - peers
      responses:
        '200':
          description: Array of peer objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Peer'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...